Add vkCmdInsertDebugUtilsLabelEXT as a frame delimiter
Currently only vkCmdDebugMarkerInsertEXT (under conditions on the
name of the marker) is considered a frame delimiter (aside from
queue present).

However, VK_EXT_debug_marker is deprecated and promoted to the
VK_EXT_debug_utils extension that support an almost identical
command vkCmdInsertDebugUtilsLabelEXT.

This commit only add this command as a frame delimiter under the
same conditions on the name of the label (which replaces the
marker).

Change-Id: Ic4164ee65eba2c1295ef8c45084ee6136dc796a8