Option for offscreen frame boundary conversion
Add `--use-ext-frame-boundary` option to convert all offscreen
frame boundaries to `VK_EXT_frame_boundary` frame boundaries.

Extensions used to specify an offscreen frame boundary can be
unsupported by the replay device. As there are many of such
extensions, it seems important to be able to "convert" these
frame boundaries to a more "standard" one.
The `VK_EXT_frame_boundary` seems to be the best choice as it
is specialy designed for this purpose, platform independent,
and supports all functionnalities supported by other extensions:
- Specifying resources attached to the frame (images and buffers)
- Depending on the execution of other commands
- Specifying frame number, label and description

Change-Id: I9fb932b219380c6f9098a16a83de2086f2b424ee