Move VSCode configuration files
Move tasks.json and launch.json to .vscode/arm subfolder.
It will never be pushed upstream anyway so we don't care about
having a special "arm" folder, and making it a subfolder allow to
not have personnal workspaced NUKED each time ou do a git pull.

In addition, minor enhancements have been made:
- Renaming of tasks: build-task, gfxrecon-capture and gfxrecon-replay
(nobody makes copies of this anyway, we just modify it as we need it)
- Default parallel jobs when compiling (16 thread looks a good default)
- Use files and directories as would do the default cmake user: build
directory is... "build" (not dbuild) and do not use "installed" files
but directly the result of the compilation.

Change-Id: I1dd1c248df31b1c9f51a3e1963201b0d45631e4f