Enable capturing fast forwarded traces to replay with rebind
This introduces a number of changes.
1. Tracking of vkCmdBuildAccelerationStructuresKHR command with new
meta command, introduced to capture the build information of the
acceleration structure even if the command buffer containing the commands
is freed before capture range.
2. Tracking of instance buffer contents of the top level acceleration structures
This buffer is used for build of the TLAS and is not guaranteed to exist when
the state recreation is written, but it is needed in the building of
TLAS, to be able to substitute the correct BLAS addresses
3. Usage of this information to facilitate replay with -m rebind
of fast forwarded traces with RT functionality

Change-Id: I3bfd2ff56123dc8010bf36bebae3bff6d043166d