Enable fast forwarding for SolarBay
- Introduce a meta command to query the compacted sizes on state recreation.

For now, each query for a compacted size is stored as a separate query. This narrows down a possibility of mistake for state recording and recreation.

The query metacommand is written before any copying meta command is written, ensuring that we get the correct sizes of when the copy is executed.

Change-Id: I43f5104c46940f6737ff10a6a6564d2fcf3c6485