Better external memory support
Until now, external memory was only supported if the platform and/or
extension(s) used where the same as for the capture device. Capturing
and replaying should be device/platform independant as much as
possible - even if device/platform dependant calls should be replayed
exactly the same if possible.

This commit tries to solve some issues to progress towards this goal.
This commit only modifies replay, not capture.

Here is what this commit does:

Minimal support for AHardwareBuffer out of Android when the following
conditions are satisfied:
- The AHardwareBuffer has a standard format with a perfect matching to
a Vulkan format
- The AHardwareBuffer data is not used (write only) OR the data that
is used is set through Vulkan commands at capture time (and not by
locking the AHardwareBuffer for CPU write)

Better support for VK_EXT_external_memory_host: Memory host pointers
should be treated in a way similar to the AHardwareBuffer, by creating
the resource first then binding it VkDeviceMemory. Doing this would
require modifications of capture commands, but this commit already
does one important change: The same pointer is used at replay time when
the same pointer was used at capture time.

Change-Id: Iec3751f7bd2c86326ced9ae62027e3e79fb4eb49