Handle portable shader group handles
With this change shader group handles are replaced before QueueSubmit
based on information from tracked vkGetRayTracingShaderGroupHandlesKHR
and vkCmdTraceRays.

This implementation assumes shader group handle size and alignment
is the same on the capture and replay device, will hit assert otherwise.
This condition is true on all tested RTU's - v0,1 and 2

Change-Id: I34a3c8f02b870cce45c9fc4a9b7bdd43576035a7