SmartGI (Underground Facility) Enabler
This commit enables the repaly of Underground Facility content.
To do this, we do CPU peek into a lot of buffers,
to find the one we need to write new device addresses too.
This is done becase the instance buffer during the build is filled
in the compute dispatch in the same command buffer.

Also, this commit adds aliases for acceleration structures.
This covers the situation when application uses
different hadles to refer to the same underlying acceleration structure.
This is specific to Underground Facility, and may be reverted later.

Change-Id: Ic7decbd0f45162f30fc36f92a109306f1180b4bb