Fix a crash on update descriptor set
It is possible that the descriptor type is not of buffer type, but the pointer to buffer infos of the write is not nullptr. It would therefore be incorrect to try to access that data in this particular write. This commit adds back the checks for descriptor type being of buffer types.

Change-Id: Ia4bbd2b7edb7d3cf8f15d14b5526b6484dc79769