Fix for SEGFAULT on replaying raytracing triangle from Khronos
The raytracing sample recreates a swapchain on each frame from the previous swapchain. Therefore, the old swapchain is destroyed, and with it, all the acquired image wrappers. The application, however, does not destroy the images when recreating the swapchain.

There was a handling for this partially in place, but it was not complete. Namely, it did not consider this scenario of recreating the swapchain from the old one.

Change-Id: If46d267b0d128057a7a13e3a4c2d198e1bf8254e