Enabling Lumen sample to be replayed with rebind
Lifting some restrictions on processing the buffers
from update descriptor sets. The issue with Lumen sample
seems to be that the compute dispatch that fills the instance buffer,
which is similar to Underground Facility, is done before the
build of top level acceleration structure is recorded.

Therefore, it was wrongfully discarded. Instead, I introduce
another restrictions based on common sense (for data transfer,
at least 2 buffers should be bound) and buffer usage flags.

Change-Id: Iefa9601c5c19e063225a2929fa51fd71b90c1b1d