Further fixes for Lumen sample content
Currently, FixDeviceAddress is not able to enable Lumen fully.
It does not work without ignoring buffers with VK_BUFFER_USAGE_INDEX_BUFFER_BIT,
which may be excluding some general purpose buffers that have a lot of usages
enabled from the update. Therefore, it cannot replace all the addresses
in this scenarios. This a fix for a fix, further improving support for
the scenario when the instance buffer is populated in compute shader,
which seems prevelant for Unreal Engine content.

Change-Id: Ib19885fb5629c03e8c7b33281a01baa9e086b58a