Rework vulkan optimizer
Currently vulkan optimizations are done in two stages separately,
using temporary files. This solution is not scalable. In preparation
for future optimizations this change introduces VulkanFileOptimizer
which similarly to Dx12FileOptimizer allows for multiple optimizations
without temporary files.

Also, rearranged optimization logic and fixed a bug when default replay
options are not set with dx12:
--set-replay-options now cannot be used with any other optimization option

Change-Id: I72c633202f3b484b9361c715c82a7104b3a50499