This commit changes the way we handle deletion of scratches.
Unfortunately, if the content does not reuse scratch buffers,
there is nothing we can really do. We cannot just erase all scratches
post queue present, because the queue present call is not blocking, and there is no
guarantee that the relevant build commands are completed.

The idea here is to introduce a double buffering approach to managing scratches.
We are deleting only the scratches from the previous queue present, and
the scratches from the current quueue present are kept alive until the next one.

This does not increase memory consumption noticeably, and it enables old
warthunder traces at the same time.

This commit also enables deletion of the internal scratches if the
application deletes the respective original buffer.

Change-Id: I11942fe1ecfe3d530584308eddc3aeff87352b6f