Handle nomali in portable raytracing
Portable raytracing mechanics relies on data retrieved from the driver.
Specifically, creating optimized AS copy requires retrieval of the compacted AS size. This data is not properly reported by nomali driver, resulting in creation of buffers of size 0. This causes unexpected out of host memory issues.

For nomali driver, it is enough to set the buffer size to any value.
In normal scenarios however, this situation should never happen,
compacted size should always be a non-zero value.

Change-Id: I3602358ef71813e147b91bace5653fe0f5cae3c3