This commit fixes the corruptions when replaying fast forwarded traces of ray tracing samples.
The root of the issue was that due to way the samples are written - i.e. destroying resources as soon as they not needed, and fast forwarding reordering some of the api calls, we ended up in the situtation when internally created instance buffer had a capture device address shared with sbt buffer, that was created later.

The solution is to modify the recorded build command to contain new address instead of the old capture address earlier then before, and put the new device address as the old device address for the buffer entry.

Change-Id: I0e53448822754c90b085db9e73c22dc04cfe394c