Fix fence query delay
When delaying fence queries, we might want to delay a
certain number of calls or for a certain number of frames.

This option allow to specify this "unit of time" for the
fence query delay option.

This commit also fix the "spam" of `vkGetFenceStatus` at
replay time by replacing the loop with a `vkWaitForFences`
call, and the `--sgfs` and `--sgfr` options not being
correctly parsed, and several other bugs related to
fence query delay.

Change-Id: Ib952203c8c492cfc228e9642536f9a3048b85047