Enable support for vkCmdUpdatebuffer and vkCmdPushConstants
This commit enables replacement of device addresses in
the data that we submit to GPU through aforementioned API calls.
This is done by utilizing FixDeviceAddress command.

This commit
also tries to keep the utilization of the FixDeviceAddress command
retrocompatible, i.e no changes to encoding / deconding the command.
However, this would mean the the id that was a memory id would be
either device id or memory id, which is determined at runtime.

Change-Id: Idf568d503979d2dba3f66fea3775732121f0a9a8