[TRACETOOLS-802] Initial implementation of the communication with the perf layer
This change enables the replayer to load the layers provided via
a replay option by name, separated by comma, and then mark the start and end of
the calls that are injected by replayer with the callbacks provided by the layers.

This change also attempts to mark as many known injected calls as possible.

This change calls this "marking layers" because the existance of performance layer is not public.

Change-Id: Ieb4d73fb0fd934a4fe20fd2371efcac6246c7414