Enabling VkQuake1
The trace we have for VkQuake1 (TRACETOOLS-933) is failing to replay as a result of trying to map a buffer that's not host visible from update descriptor set command. Since that buffer is not the one that contains any relevant data (we go through a batch of buffers that 'may' be relevant) then skipping buffers we can't map is more or less the only option we have at the moment. Skipping the buffers that are in non host visible memories enables the content VkQuake1 to be replayed.

Change-Id: Ie2958eb23d8f99d8e12dbf34fad41486f93ca022