Fix for wrongfull deleting of image wrapper in state tracker
State tracker, when destroying the state for
the swapchain, did not take into account the possible scanerio of
swapchain recreation using the old swapchain. This means that the
swapchain images will be reused. Stete tracker was destroying the image wrappers.

This commit tries to fix this, by not destroing the image wrappers for
the images that have >= 1 parent swapchain. However, not all
the images from the old swapchain are reused. Therefore, we need to add
a post process, that will ensure that only the images that
are used for both swapchains have several parents.

We also update the create_parameters on insertion of duplicate handle,
to make sure that we have a vkGetSwapchainImagesKHR with the alive swapchain
in the state recreation.

Change-Id: I524c0b2b453963eda2fc6d63b95d4cecca5433ac