Unified buffer recreation for rebuilding acceleration structures on fastforwarding
To make fastforwarding portable, we need to actually rebuild acceleration structures
in state recreation. This may be complicated, as some of the buffers may be
not present in state recreation by default, on account of them being destroyed.

This commit tries to change that. Namely, it trackes the buffers,
that were used as inputs to acceleration structure building.
If these buffers end up destroyed, these changes make sure that those
buffers are created and initialized with correct data nevertheless,
and destroyed right after they are not needed anymore, to not pollute the trace.

The caveat is that these changes may introduce a bigger memory footprint during
capture, as we need to preserve a copy of the data when the buffer gets destroyed.
To get the buffers a distinct device address from those that may be reused, we
issue mock device addresses

These changes should not break already existing fastforwarding traces.

Change-Id: I6a3963f0b3d2d70747f4d668760170c547bc7a8e