Save InitBuffer commands required for builds
Optimizer may remove the resources if they are not
added to a user, which is a command buffer.
For micromap fastforwarding, as it creates
it's own command buffer, this works almost out of the box,
by adding the processing of the build command.

Metacommands, on the other hand, don't have a command buffer
tied to them. Therefore, for build metacommand, the buffers
are unconditionally preserved. This is not a good solution,
but a fast one. Better solution is to remove the build metacommands
alltogether and use the approach devised for micromap fastforwarding.

Change-Id: I4982835358c81b56e5f65fa82e47bf4b09bbb5b7