Fix `ArgumentParser::SplitStringByFlag`
The current function splits the string according ot the given
flag but also tries to get rid of the surrounding spaces.
There is also a bug when the string is a single character
surrounded (or not) by spaces, as in this case `begin != end`
is `false` and the character is ignored.

I propose to stop trying to remove surrounding spaces, as the
user is responsible for the spaces he provides to the replayer.

Change-Id: I5abed0c44c78bfecccd8c2932def88fbdd5aacf4