Fix layer communication and instance creation
- The capture layer supports an instance extension (`VK_EXT_debug_utils`)
but this extension was not "advertised" by the layer's
`vkEnumerateInstanceExtensionProperties` implementation
- The capture layer was using the internal
`vkEnumerateInstanceExtensionProperties` to check if an extension is
supported by the underlying layer/driver or not. This is a mistake as this
internal version of the function only shows extensions supported by the
layer itself
- When the layer is faking `VK_EXT_debug_utils` and that the application
created a `VkDebugUtilsMessengerEXT`, the capture layer didn't set the
`pMessenger` variable to a non-null value, this is incorrect behavior
according to Vulkan and in the particular case of the application being
the replayer, the corresponding wrapper was not created, leading to a
crash when the replayer would call `vkDestroyDebugUtilsMessengerEXT
- At replay time, support for `VK_EXT_debug_utils` was not properly
checked to determine if it would be actually supported or faked.

Change-Id: I9f11dcb89727bbae85cab4557527b040cb1bb2fd