Add `--screenshot-interval` option
In some case people want to be able to take a screenshot "every N frame"
over long frame ranges or over the entire capture.

For now the only way to do this was either to take all screenshots and
discard unused ones afterwards - which increase significantly the run
time - or to write a very long list manually or assisted with a script,
which is not very user-friendly.

This commit implements this with an option `--screenshot-interval N`
where a screenshot is taken every `N` frame within screenshot ranges.

The default behavior stays the same if `--screenshot-interval` is `1`,
which is the default value.

Example: If the screenshot ranges are `10-15,20-30` and the screenshot
interval is `3`, then screenshots will be taken at frames 10, 13,
20, 23, 26 and 29.

Change-Id: I5072081febb626477e659cbe4089e79b7d8075eb