Check for changes before starting auto-merge
To dodge the "Oopsie" situation of running the script
and seeing the work of a week gone to ashes.

This commit also make the `-b` argument required to
force the user to be reminded the branch he's on and
the branch he wants to merge

It also add a print of the total number of commits
to merge even if none could be merged.

Change-Id: I5453861136a7fc42d7f10f8fa5808dbfa6d83278