Fix CMake unit tests build directory
Currently `test.py` build directory is "guessed" using the
architecture and configuration of the build, assuming the
build is done using `build.py`.

This is an unecessarry and unsafe assumption, as the build
could be started directly from a CMake command.

This commit fixes that by specifying the build directory
when launching `test.py` using CMake's
`CMAKE_CURRENT_BINARY_DIR`. This will work no matter how the
CMake script is started.

Change-Id: I813c2745560a0d74a53f4da8c96044e4d3d8e5bb