Workaround for stray calls from VK_KHR_get_surface_capabilities2
Until https://github.com/LunarG/gfxreconstruct/pull/2265, VulkanStateWriter is always using VK_KHR_get_surface_capabilities2 when writing state, even if the original trace / application did not request the extension.

Even though the original bug is fixed, there are traces in the wild with this behavior, which on some platforms results in a crash. This change enables the extension if the replay device supports the extension, even if the trace does not request it.

Change-Id: I738b8c177c62534de35b6b9dccd6f2f2c7380e03